+++ community/src/packages/views/mantine/src/components/MtCheckboxGroup.tsx
@@ -64,7 +64,7 @@ export const mtCheckboxGroup = define(MtCheckboxGroup, 'MtCheckboxGroup')
     description: description,
     error: string,
     value: array.valued,
-    items: array
+    items: array.localize
       .default(toLabeledValues(['Option 1', 'Option 2', 'Option 3'])),
     variant: filledVariant,
     size: size,