+++ community/src/packages/views/mantine/src/components/MtNativeSelect.tsx
@@ -10,7 +10,7 @@ export const mtNativeSelect = define(NativeSelect, 'MtNativeSelect')
   .props({
     ...baseInputProps,
     ...inputSectionProps,
-    data: array
+    data: array.localize
       .default(toLabeledValues(['Option 1', 'Option 2', 'Option 3'])),
   })
   .overrideEventHandlers(valueEventHandlers)