+++ community/src/packages/views/mantine/src/components/MtSwitchGroup.tsx
@@ -53,7 +53,7 @@ export const mtSwitchGroup = define(MtSwitchGroup, 'MtSwitchGroup')
     description: description,
     error: string,
     value: array.valued,
-    items: array
+    items: array.localize
       .default(toLabeledValues(['Option 1', 'Option 2', 'Option 3'])),
     size: size,
     labelPosition: labelPosition,